move models/metaplex to common pkg
