+++ js/packages/common/src/contexts/index.tsx
@@ -4,5 +4,4 @@ export * as Connection from './connection';
 export * from './connection';
 export * as Wallet from './wallet';
 export * from './wallet';
-export * as Store from './store';
 export * from './store';
