+++ js/packages/common/src/models/metaplex/deprecatedInitAuctionManagerV1.ts
@@ -1,4 +1,3 @@
-import { programIds, StringPublicKey, toPublicKey } from '@oyster/common';
 import {
   SystemProgram,
   SYSVAR_RENT_PUBKEY,
@@ -7,6 +6,7 @@ import {
 import { serialize } from 'borsh';
 
 import { getAuctionKeys, SCHEMA } from '.';
+import { programIds, StringPublicKey, toPublicKey } from '../../utils';
 import {
   AuctionManagerSettingsV1,
   DeprecatedInitAuctionManagerV1Args,
