+++ js/packages/web/src/actions/createAuctionManager.ts
@@ -5,7 +5,6 @@ import {
   SystemProgram,
 } from '@solana/web3.js';
 import {
-  actions,
   Metadata,
   ParsedAccount,
   MasterEditionV1,
@@ -41,7 +40,8 @@ import {
   TupleNumericType,
   SafetyDepositConfig,
   ParticipationStateV2,
-} from '../models/metaplex';
+} from '@oyster/common/dist/lib/models/metaplex/index';
+import { createTokenAccount } from '@oyster/common/dist/lib/actions/account';
 import { createVault } from './createVault';
 import { closeVault } from './closeVault';
 import {
@@ -50,15 +50,13 @@ import {
 } from './addTokensToVault';
 import { makeAuction } from './makeAuction';
 import { createExternalPriceAccount } from './createExternalPriceAccount';
-import { deprecatedValidateParticipation } from '../models/metaplex/deprecatedValidateParticipation';
+import { deprecatedValidateParticipation } from '@oyster/common/dist/lib/models/metaplex/deprecatedValidateParticipation';
 import { deprecatedCreateReservationListForTokens } from './deprecatedCreateReservationListsForTokens';
 import { deprecatedPopulatePrintingTokens } from './deprecatedPopulatePrintingTokens';
 import { setVaultAndAuctionAuthorities } from './setVaultAndAuctionAuthorities';
 import { markItemsThatArentMineAsSold } from './markItemsThatArentMineAsSold';
-import { validateSafetyDepositBoxV2 } from '../models/metaplex/validateSafetyDepositBoxV2';
-import { initAuctionManagerV2 } from '../models/metaplex/initAuctionManagerV2';
-
-const { createTokenAccount } = actions;
+import { validateSafetyDepositBoxV2 } from '@oyster/common/dist/lib/models/metaplex/validateSafetyDepositBoxV2';
+import { initAuctionManagerV2 } from '@oyster/common/dist/lib/models/metaplex/initAuctionManagerV2';
 
 interface normalPattern {
   instructions: TransactionInstruction[];
