+++ js/packages/web/src/actions/deprecatedCreateReservationListsForTokens.ts
@@ -6,7 +6,7 @@ import {
 } from '@oyster/common';
 import { WalletNotConnectedError } from '@solana/wallet-adapter-base';
 import { SafetyDepositInstructionTemplate } from './addTokensToVault';
-import { WinningConfigType } from '../models/metaplex';
+import { WinningConfigType } from '@oyster/common/dist/lib/models/metaplex/index';
 
 const BATCH_SIZE = 10;
 // This command batches out creating reservation lists for those tokens who are being sold in PrintingV1 mode.
