+++ js/packages/web/src/actions/saveAdmin.ts
@@ -5,10 +5,10 @@ import {
   sendTransactionWithRetry,
   WalletSigner,
 } from '@oyster/common';
+import { WhitelistedCreator } from '@oyster/common/dist/lib/models/metaplex/index';
+import { setStore } from '@oyster/common/dist/lib/models/metaplex/setStore';
+import { setWhitelistedCreator } from '@oyster/common/dist/lib/models/metaplex/setWhitelistedCreator';
 import { WalletNotConnectedError } from '@solana/wallet-adapter-base';
-import { WhitelistedCreator } from '../models/metaplex';
-import { setStore } from '../models/metaplex/setStore';
-import { setWhitelistedCreator } from '../models/metaplex/setWhitelistedCreator';
 
 // TODO if this becomes very slow move to batching txns like we do with settle.ts
 // but given how little this should be used keep it simple
