+++ js/packages/web/src/hooks/useAuctions.ts
@@ -18,14 +18,15 @@ import { useMeta } from '../contexts';
 import {
   AuctionManager,
   AuctionManagerStatus,
+  AuctionManagerV1,
   AuctionManagerV2,
   BidRedemptionTicket,
   BidRedemptionTicketV2,
   getBidderKeys,
   SafetyDepositConfig,
   WinningConfigType,
-} from '../models/metaplex';
-import { AuctionManagerV1 } from '../models/metaplex/deprecatedStates';
+  AuctionViewItem,
+} from '@oyster/common/dist/lib/models/metaplex/index';
 
 export enum AuctionViewState {
   Live = '0',
@@ -33,14 +34,6 @@ export enum AuctionViewState {
   Ended = '2',
   BuyNow = '3',
   Defective = '-1',
-}
-
-export interface AuctionViewItem {
-  winningConfigType: WinningConfigType;
-  amount: BN;
-  metadata: ParsedAccount<Metadata>;
-  safetyDeposit: ParsedAccount<SafetyDepositBox>;
-  masterEdition?: ParsedAccount<MasterEditionV1 | MasterEditionV2>;
 }
 
 // Flattened surface item for easy display
