+++ js/packages/web/src/views/admin/index.tsx
@@ -12,7 +12,7 @@ import {
   Divider,
 } from 'antd';
 import { useMeta } from '../../contexts';
-import { Store, WhitelistedCreator } from '../../models/metaplex';
+import { Store, WhitelistedCreator } from '@oyster/common/dist/lib/models/metaplex/index';
 import {
   MasterEditionV1,
   notify,
