+++ js/packages/web/src/views/home/setup.tsx
@@ -2,7 +2,7 @@ import {
   useConnection,
   useStore,
   useWalletModal,
-  WalletSigner,
+  WhitelistedCreator,
 } from '@oyster/common';
 import { useWallet } from '@solana/wallet-adapter-react';
 import { Button } from 'antd';
@@ -10,9 +10,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { saveAdmin } from '../../actions/saveAdmin';
 import { useMeta } from '../../contexts';
-import { WhitelistedCreator } from '../../models/metaplex';
 import { SetupVariables } from '../../components/SetupVariables';
-import { WalletAdapter } from '@solana/wallet-adapter-base';
 
 export const SetupView = () => {
   const [isInitalizingStore, setIsInitalizingStore] = useState(false);
